Separate progress bar code into a dedicated file for now
