+++ downloader/downloader.go
@@ -34,17 +34,6 @@ var (
 	noResume bool
 )
 
-// progress bar elements
-var (
-	progressIndicator *uilive.Writer
-	progressStatus    string
-
-	timeoutCount int
-	errorCount   int
-
-	averageTimePer1000 float64 = 1
-)
-
 type Resumer struct {
 	OutputFilename string `json:"outputFilename"`
 
@@ -432,6 +421,7 @@ func (r *Resumer) nextChunkNumber() (nextChunkNumber, skipNRows uint64) {
 	if r.DoneElements == 0 {
 		nextChunkNumber = 0
 		skipNRows = 0
+		client.SetNextChunkNumber(0)
 		return
 	}
 
@@ -616,8 +606,8 @@ func totalElements() (uint64, error) {
 // containing the total number of elements.
 func (r *Resumer) fetchTotalElements() ([]byte, int, error) {
 
-	path := fmt.Sprintf("/2.0/crawls/%v/%s", client.CrawlID, client.Mode)
-	method := "GET"
+	path := client.GetRelativePath()
+	method := client.GetRequestMethod()
 
 	headers := http.Header{}
 	bodyParameters := url.Values{}
@@ -660,45 +650,4 @@ func (r *Resumer) PersistConfig() error {
 		return err
 	}
 	return nil
-}
-
-// progress outputs the progress percentage
-func (r *Resumer) progress() *big.Float {
-	var progressPerc *big.Float = big.NewFloat(0.0)
-	if res.TotalElements > 0 && res.DoneElements > 0 {
-		progressPerc = big.NewFloat(0).Quo(big.NewFloat(100), big.NewFloat(0).Quo(big.NewFloat(0).SetUint64(res.TotalElements), big.NewFloat(0).SetUint64(res.DoneElements)))
-	}
-	return progressPerc
-}
-
-// updateStatus sets the first part of the progress bar message
-func updateStatus(s string) {
-	// TODO: add a mutex?
-	progressStatus = s
-}
-
-// progress animation
-func progressLoop() {
-	var n int = 0
-	var max int = 10
-	for {
-
-		ETAuint64, _ := big.NewFloat(0).Quo(big.NewFloat(0).Quo(big.NewFloat(0).Sub(big.NewFloat(0).SetUint64(res.TotalElements), big.NewFloat(0).SetUint64(res.DoneElements)), big.NewFloat(1000)), big.NewFloat(averageTimePer1000)).Uint64()
-		ETAtime := time.Duration(ETAuint64) * time.Millisecond * 110
-		ETAstring := ETAtime.String()
-
-		progressMessage := progressStatus + chs(n, ".") + chs(max-n, "*")
-		progressMessage = progressMessage + fmt.Sprintf(" | ETA %v |", ETAstring)
-		progressMessage = progressMessage + fmt.Sprintf(" Chunk size %v |", res.chunkSize)
-		progressMessage = progressMessage + fmt.Sprintf(" %v timeouts |", timeoutCount)
-		progressMessage = progressMessage + fmt.Sprintf(" %v errors |", errorCount)
-
-		fmt.Fprintln(progressIndicator, progressMessage)
-		time.Sleep(time.Millisecond * 500)
-
-		n++
-		if n >= max {
-			n = 0
-		}
-	}
 }
